Optimize API functions for Vercel deployment limit
- Consolidate 8 individual API functions into 2 consolidated services
- Create api/ai-services.py combining 6 AI-related endpoints
- Create api/materials-services.py combining materials and quiz generation
- Reduce from 13 to 7 API functions (under 12 function limit)
- Update frontend components to use new consolidated API endpoints
- Maintain all functionality while reducing deployment complexity
- Fix Vercel Hobby plan function count limit issue
